+++ database/migrations/2021_05_14_000851_create_d_e_o_s_controllers_table.php
@@ -13,20 +13,37 @@ class CreateDEOSControllersTable extends Migration
      */
     public function up()
     {
-        Schema::create('deos_controllers', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('deos_controllers')) {
+            Schema::create('deos_controllers', function (Blueprint $table) {
+                $table->id();
+                $table->text('name')->nullable();
+                $table->string('ip_address')->nullable();
+                $table->string('port_number')->nullable();
+                $table->double('longitude', 10, 4)->nullable();
+                $table->double('latitude', 10, 4)->nullable();
+                $table->boolean('enable_weather_forecast')->nullable();
+                $table->boolean('enable_electricityprice_forecast')->nullable();
 
-            $table->text('name')->nullable();
-            $table->string('ip_address')->nullable();
-            $table->string('port_number')->nullable();
-            $table->double('longitude', 10, 4)->nullable();
-            $table->double('latitude', 10, 4)->nullable();
-            $table->boolean('enable_weather_forecast')->nullable();
-            $table->boolean('enable_electricityprice_forecast')->nullable();
+                $table->unsignedBigInteger('building_id')->nullable();
+                $table->foreign('building_id')->references('id')->on('buildings')->onDelete('cascade');
+                $table->timestamps();
+            });
+        } else
+        Schema::table('deos_controllers', function (Blueprint $table) {
+            if (!Schema::hasColumn('deos_controllers', 'id')) {
+                $table->id();
+                $table->text('name')->nullable();
+                $table->string('ip_address')->nullable();
+                $table->string('port_number')->nullable();
+                $table->double('longitude', 10, 4)->nullable();
+                $table->double('latitude', 10, 4)->nullable();
+                $table->boolean('enable_weather_forecast')->nullable();
+                $table->boolean('enable_electricityprice_forecast')->nullable();
 
-            $table->unsignedBigInteger('building_id')->nullable();
-            $table->foreign('building_id')->references('id')->on('buildings')->onDelete('cascade');
-            $table->timestamps();
+                $table->unsignedBigInteger('building_id')->nullable();
+                $table->foreign('building_id')->references('id')->on('buildings')->onDelete('cascade');
+                $table->timestamps();
+            }
         });
     }
