+++ database/migrations/2021_09_01_203416_create_csv__trend__data_table.php
@@ -13,14 +13,27 @@ class CreateCsvTrendDataTable extends Migration
      */
     public function up()
     {
-        Schema::create('csv_trend_data', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('trend_group_id');
-            $table->foreign('trend_group_id')->references('id')->on('trend_groups')->onDelete('cascade');
-            $table->timestamp('timestamp')->nullable();
-            $table->string('sensor_name')->nullable();
-            $table->float('sensor_value')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('csv_trend_data')) {
+            Schema::create('csv_trend_data', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('trend_group_id');
+                $table->foreign('trend_group_id')->references('id')->on('trend_groups')->onDelete('cascade');
+                $table->timestamp('timestamp')->nullable();
+                $table->string('sensor_name')->nullable();
+                $table->float('sensor_value')->nullable();
+                $table->timestamps();
+            });
+        } else
+        Schema::table('csv_trend_data', function (Blueprint $table) {
+            if (!Schema::hasColumn('csv_trend_data', 'id')) {
+                $table->id();
+                $table->unsignedBigInteger('trend_group_id');
+                $table->foreign('trend_group_id')->references('id')->on('trend_groups')->onDelete('cascade');
+                $table->timestamp('timestamp')->nullable();
+                $table->string('sensor_name')->nullable();
+                $table->float('sensor_value')->nullable();
+                $table->timestamps();
+            }
         });
     }
